SW-13 Refactor: change file name conventions
