+++ apps/web/app/(admin)/admin/page.tsx
@@ -1,6 +1,6 @@
 import { NextPage } from 'next';
 import dynamic from 'next/dynamic';
-const AdminApp = dynamic(() => import('./_components/AdminApp'), { ssr: false });
+const AdminApp = dynamic(() => import('./_components/admin-app'), { ssr: false });
 
 const Page: NextPage = () => <AdminApp />;
