+++ apps/web/app/(admin)/admin/posts/_components/extraInfo/class-score-input.tsx
