+++ apps/web/app/(external)/_components/Header/desktop-nav-menu.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 
 import { nav_menu } from '@/libs/data/nav_menu';
 
-import { ThemeToggle } from './ThemeToggle';
+import { ThemeToggle } from './theme-toggle';
 
 function DesktopHeaderNavMenu() {
   return (
