+++ apps/web/app/(external)/_components/sections/Intro/speech-bubble.tsx
