+++ apps/web/app/(external)/_components/sections/Process/_components/arrow.tsx
