+++ apps/web/app/(external)/_components/sections/_components/section-title.tsx
