+++ apps/web/app/(external)/layout.tsx
@@ -6,9 +6,9 @@ import { cn } from '@repo/ui/lib/utils';
 
 import { noto_sans_kr } from '@/libs/font';
 
-import Header from './_components/Header';
-import QueryProvider from './_components/providers/QueryProvider';
-import { ThemeProvider } from './_components/providers/ThemeProvider';
+import Header from './_components/header';
+import QueryProvider from './_components/providers/query-provider';
+import { ThemeProvider } from './_components/providers/theme-provider';
 
 export const metadata: Metadata = {
   title: 'Santosha Yoga Studio',
